refactor: rename loading to loadingOverlay
